[test] Introduce Docker based test invorker (#18)
*Motivation*

when developing hive storage handler, we need to run integration
tests for testing storage handler reading data directly from bookkeeper.
this requires the test code is running in the same network where
bookies are running.

since the integration test is written in testcontainers, all bookies
are deployed in containers. it is non-trivial to expose bookies'
addresses to allow both test codes and broker to access.

so ideally we should support running the tests in a testcontainer
where is running in the same network as other containers.

*Modifications*

Implement a Docker (via testcontainer) based TestInvoker.
The docker-based TestInvoker will do followings when executing a test method:

- create a `java:8` testcontainer
- mount the module uber jar as `/test/dockertest.jar`
- launch the testcontainer to invoke a specific test in the container